- typo fixed - types reorganized
